Added surefire plugin and refactored to write some tests.
Wrestled with LEB128 a while trying to understand exactly what was
going on. Eventually settled on VarInt being an unsigned LEB128
which provided consistent results with the wiki and the existing
implementation. Converted the handshake packet into its own class
to test it independently of Bot. Added some other boilerplate like
the Packet interface to enable plumbing like the PacketDataEncoder
to simplify the code in Bot to make it more readable. One last
oddity to point out. It would seem that converting an int to a
short simply drops the 16 highest order bits effectively making it
an unsigned conversion. This is why you can convert 65535 to a
short, which has a positive limit of 32767, and still get the
right sequence of bytes in the output. Java would understand the
value to actually be -1, however, so it's important not to try to
do math after such a cast or you will get unexpected results.

Wiki: https://wiki.vg/Protocol#VarInt_and_VarLong